requirements.txt
flask==2.2.5
google-generativeai==0.3.1
python-dotenv==1.0.0 